+++ src/AeroGA/AeroGA.py
@@ -1,18 +1,17 @@
-from datetime import datetime
+import copy
 import time
 import random
-import numpy as np
-import pandas as pd
-import math
 import multiprocessing
 from statistics import mean
-from bisect import bisect_left
-import matplotlib.pyplot as plt
-import plotly.express as px
 from . import settings
-from .classes import Individual, ErrorType, Log
-import os
-import copy
+from .Classes import Individual, ErrorType, Log
+from .Operators_Selection import roulette_selection, tournament_selection, rank_selection
+from .Operators_Crossover import arithmetic_crossover, SBX_crossover, crossover_1pt, crossover_2pt
+from .Operators_Mutation import gaussian_mutation, polynomial_mutation, online_expected_diversity, online_parameter
+from .Population import denormalize_population, denormalize_individual, generate_population_normalized
+from .Evaluators import parallel_fitness, fitness
+from .Performance_Metrics import diversity_metric
+from .Utilities import export_excell, create_plotfit, create_boxplots, parallel_coordinates
 
 # Setting error log file
 ErrorLog = Log("error.log")
@@ -24,9 +23,10 @@ ErrorLog = Log("error.log")
 def optimize(selection = "tournament", crossover = "1-point", mutation = "gaussian", n_threads = -1,
     min_values = list, max_values = list, num_variables = int, num_generations = int, elite_count = int, elite="local",
     plotfit = True, plotbox = False, plotparallel = False, TabuList = False, penalization_list = [1000],
-    fitness_fn = None, classe = "default"
-    ):
+    fitness_fn = None, classe = "default"):
+
     """Perform the genetic algorithm to find an optimal solution."""
+
     try:
         t_inicial = time.time()
 
@@ -310,618 +310,3 @@ def optimize(selection = "tournament", crossover = "1-point", mutation = "gaussi
         return ErrorType("danger", str(e))
 
     return out
-
-# #####################################################################################
-# ##################################### Fitness #######################################
-# #####################################################################################
-
-def parallel_fitness(population = list, fitness_fn = None, num_processes = int):
-    """Calculate the fitness of each individual in the population."""
-    with multiprocessing.Pool(num_processes) as pool:
-        fitness_values = pool.map(fitness_fn, population)
-    return fitness_values
-
-# Fitness function without using multi threads
-def fitness(population = list, fitness_fn = None):
-    """Calculate the fitness of each individual in the population."""
-    return [fitness_fn(ind) for ind in population]   
-
-# #####################################################################################
-# #################################### Init Pop #######################################
-# #####################################################################################
-
-# define the genetic algorithm functions
-def generate_population(size = int, num_variables = int, min_values = list, max_values = list):
-    """Generate a population of random genes."""
-    population = [[round(random.uniform(min_values[i], max_values[i]),4) if isinstance(min_values[i],float) else random.randint(min_values[i], max_values[i]) for i in range(num_variables)] for _ in range(size)]
-    return population
-
-def generate_population_normalized(size = int, num_variables = int, min_values = list, max_values = list):
-    """Generate a population of normalized random genes."""
-    population = []
-    for _ in range(size):
-        individual = []
-        for i in range(num_variables):
-            if isinstance(min_values[i], float):
-                gene = random.uniform(min_values[i], max_values[i])
-            else:
-                gene = random.randint(min_values[i], max_values[i])
-            
-            if min_values[i] == max_values[i]:
-                normalized_gene = min_values[i]  # Atribui o valor mínimo diretamente
-            else:
-                normalized_gene = (gene - min_values[i]) / (max_values[i] - min_values[i])
-            
-            individual.append(round(normalized_gene, 4))
-        
-        population.append(individual)
-    
-    return population
-
-# #####################################################################################
-# ################################### Selection #######################################
-# #####################################################################################
-
-def roulette_selection(population = list, fitness_values = list):
-    """Select two parents using roulette wheel selection."""
-    total_fitness = sum(fitness_values)
-    pick = random.uniform(0, 1/total_fitness)
-
-    current = 0
-    for i, ind in enumerate(population):
-        current += 1/fitness_values[i]
-        if current > pick:
-            parent = ind
-            break
-    
-    return parent
-
-def tournament_selection(population = list, fitness_values = list, tournament_size = int):
-    """Select two parents using tournament selection."""
-    if len(population) < tournament_size:
-        tournament_pop = population
-    else:
-        tournament_pop = random.sample(population, tournament_size)
-    tournament_fitness = [fitness_values[population.index(ind)] for ind in tournament_pop]
-    parent = tournament_pop[tournament_fitness.index(min(tournament_fitness))]
-
-    return parent
-
-def rank_selection(population = list, fitness_values = list):
-    """Select two parents using rank selection."""
-    n = len(population)
-    fitness_ranks = list(reversed(sorted(range(1, n+1), key=lambda x: fitness_values[x-1])))
-    cumulative_prob = [sum(fitness_ranks[:i+1])/sum(fitness_ranks) for i in range(n)]
-    parent = population[bisect_left(cumulative_prob, random.random())]
-
-    return parent
-
-# #####################################################################################
-# ################################### Crossover #######################################
-# #####################################################################################
-
-def arithmetic_crossover(parent1 = list, parent2 = list, min_values = list, max_values = list, alpha = 0.05):
-    """Apply arithmetic crossover to produce two offspring."""
-    offspring1 = []
-    offspring2 = []
-    for i in range(len(parent1)):
-        offspring1.append(round(alpha*parent1[i] + (1-alpha)*parent2[i],4))
-        offspring2.append(round(alpha*parent2[i] + (1-alpha)*parent1[i],4))
-        if not isinstance(parent1[i], float):
-            offspring1[i] = int(offspring1[i])
-            offspring2[i] = int(offspring2[i])
-
-        # Ensure the offspring stay within the bounds
-        offspring1[i] = min(max(offspring1[i], min_values[i]), max_values[i])
-        offspring2[i] = min(max(offspring2[i], min_values[i]), max_values[i])
-
-    return offspring1, offspring2
-
-def SBX_crossover(parent1 = list, parent2 = list, min_values = list, max_values = list, eta=0.5):
-    """Apply SBX crossover to produce two offspring."""
-    # Generate a random number for each variable
-    u = random.uniform(0, 1)
-    
-    # Calculate the beta value for each variable
-    beta = 0
-    if u <= 0.5:
-        beta = (2 * u) ** (1 / (eta + 1))
-    else:
-        beta = (1 / (2 * (1 - u))) ** (1 / (eta + 1))
-    
-    # Calculate the offspring for each variable
-    offspring1 = []
-    offspring2 = []
-    for i in range(len(parent1)):
-        offspring1.append(round(0.5 * ((1 + beta) * parent1[i] + (1 - beta) * parent2[i]),4))
-        offspring2.append(round(0.5 * ((1 - beta) * parent1[i] + (1 + beta) * parent2[i]),4))
-        
-        # Ensure the offspring stay within the bounds
-        offspring1[i] = min(max(offspring1[i], min_values[i]), max_values[i])
-        offspring2[i] = min(max(offspring2[i], min_values[i]), max_values[i])
-        
-        # Round integer values to the nearest integer
-        if isinstance(parent1[i], int):
-            offspring1[i] = round(offspring1[i])
-            offspring2[i] = round(offspring2[i])
-    
-    return offspring1, offspring2
-
-def crossover_1pt(parent1 = list, parent2 = list):
-    """Apply 1 Point crossover to produce two offspring."""
-    n = len(parent1)
-    cxpoint = random.randint(1, n-1)
-    offspring1 = parent1[:cxpoint] + parent2[cxpoint:]
-    offspring2 = parent2[:cxpoint] + parent1[cxpoint:]
-    return offspring1, offspring2
-
-def crossover_2pt(parent1 = list, parent2 = list):
-    """Apply 2 Point crossover to produce two offspring."""
-    size = len(parent1)
-    cxpoint1, cxpoint2 = sorted(random.sample(range(size), 2))
-    parent1[cxpoint1:cxpoint2], parent2[cxpoint1:cxpoint2] = parent2[cxpoint1:cxpoint2], parent1[cxpoint1:cxpoint2]
-    return parent1, parent2
-
-
-# #####################################################################################
-# #################################### Mutation #######################################
-# #####################################################################################
-
-def gaussian_mutation(individual = list, min_values = list, max_values = list, std_dev = float):
-    """Perform gaussian mutation on an individual."""
-    mutated_genes = []
-    for i in range(len(individual)):
-        if isinstance(individual[i], int):
-            mutated_gene = min(max(round(random.gauss(individual[i], std_dev)),min_values[i]), max_values[i])
-        else:
-            mutated_gene = round(min(max(random.gauss(individual[i], std_dev),min_values[i]), max_values[i]),4)
-        mutated_genes.append(mutated_gene)
-    return mutated_genes
-
-def polynomial_mutation(individual = list, min_values = list, max_values = list, eta = int):
-    """Perform polynomial mutation on an individual."""
-    mutated_genes = []
-    for i in range(len(individual)):
-        if isinstance(individual[i], int):
-            if max_values[i] != min_values[i]:
-                if ((individual[i] - min_values[i]) / (max_values[i] - min_values[i])) < 0.5:
-                    delta = (2 * (individual[i] - min_values[i]) / (max_values[i] - min_values[i])) ** (1 + (eta + random.uniform(0,1))) - 1
-                else:
-                    delta = 1 - (2 * (max_values[i] - individual[i]) / (max_values[i] - min_values[i])) ** (1 + (eta + random.uniform(0,1)))
-                mutated_gene = round(max(min_values[i], min(individual[i] + delta, max_values[i])))
-            else:
-                if ((individual[i] - min_values[i])) < 0.5:
-                    delta = 0
-                else:
-                    delta = 0
-                mutated_gene = round(max(min_values[i], min(individual[i] + delta, max_values[i])))
-        else:
-            if max_values[i] != min_values[i]:
-                if ((individual[i] - min_values[i]) / (max_values[i] - min_values[i])) < 0.5:
-                    delta = (2 * (individual[i] - min_values[i]) / (max_values[i] - min_values[i])) ** (1 + (eta + random.uniform(0,1))) - 1
-                else:
-                    delta = 1 - (2 * (max_values[i] - individual[i]) / (max_values[i] - min_values[i])) ** (1 + (eta + random.uniform(0,1)))
-                mutated_gene = round(max(min_values[i], min(individual[i] + delta, max_values[i])),4)
-            else:
-                if ((individual[i] - min_values[i])) < 0.5:
-                    delta = 0
-                else:
-                    delta = 0
-                mutated_gene = round(max(min_values[i], min(individual[i] + delta, max_values[i])),4)
-        mutated_genes.append(mutated_gene)
-    return mutated_genes
-
-# #####################################################################################
-# ################################## Normalization ####################################
-# #####################################################################################
-
-def denormalize_individual(individual = list, min_values = list, max_values = list, classe = str):
-    """Denormalize the genes of an individual."""
-    denormalized_individual = []
-    for i in range(len(individual)):
-        gene = individual[i]
-        min_val = min_values[i]
-        max_val = max_values[i]
-
-        if classe == "default" or "Micro":
-            rounding_value = 4
-        elif classe == "Regular":
-            rounding_value = 6
-
-        if isinstance(min_val, int) and isinstance(max_val, int):
-            denormalized_gene = round((gene * (max_val - min_val)) + min_val)
-        else:
-            denormalized_gene = round((gene * (max_val - min_val)) + min_val, rounding_value)
-
-        denormalized_individual.append(denormalized_gene)
-
-    return denormalized_individual
-
-
-def denormalize_population(population = list, min_values = list, max_values = list, classe = str):
-    """Denormalize the entire population."""
-    denormalized_population = []
-    for individual in population:
-        denormalized_individual = denormalize_individual(individual, min_values, max_values, classe)
-        denormalized_population.append(denormalized_individual)
-    
-    return denormalized_population
-
-# #####################################################################################
-# #################################### Metrics ########################################
-# #####################################################################################
-
-def diversity_metric(population = list):
-    """Calculate the normalized average distance between individuals in the population."""
-    
-    if not population or len(population) < 2:
-        return 0.0
-
-    total_distance = 0
-    max_possible_distance = 0
-
-    for i in range(len(population)):
-        for j in range(i + 1, len(population)):
-            ind1 = Individual(population[i])
-            ind2 = Individual(population[j])
-            total_distance += sum(abs(ind1.genes[k] - ind2.genes[k]) for k in range(len(ind1.genes)))
-            max_possible_distance += max(abs(ind1.genes[k] - ind2.genes[k]) for k in range(len(ind1.genes)))
-
-    # Calculate normalized average distance
-    if max_possible_distance == 0:
-        normalized_average_distance = 0
-    else:
-        normalized_average_distance = total_distance / ((len(population[0])/2.45)*max_possible_distance)
-
-    return round(normalized_average_distance, 4)
-
-# #####################################################################################
-# ################################ Online Parameters ##################################
-# #####################################################################################
-
-def online_parameter(online_control = bool, num_generations = int, mutation_prob = float, crossover_prob = float):
-    """Calculate the probability for crossover and mutation each generation, the values respscts a exponencial function, that for mutation
-       decreases each generation and increases for crossover. If online control is False than it is used the fixed parameters. 
-    
-        # MUTPB_LIST: Mutation Probability
-        # CXPB_LIST: Crossover Probability
-    """
-
-    if online_control == True:
-        line_x = np.linspace(start=1, stop=50, num=num_generations)
-        MUTPB_LIST = (-(np.log10(line_x) - np.log10(line_x[0]))/(np.log10(line_x[-1]) - np.log10(line_x[0])) + 1) * mutation_prob
-        
-        # line_x = np.linspace(start=1, stop=5, num=num_generations)
-        # CXPB_LIST = (np.log10(line_x) - np.log10(line_x[0]))/(np.log10(line_x[-1]) - np.log10(line_x[0]))
-        CXPB_LIST = [crossover_prob]*num_generations
-    else:
-        MUTPB_LIST = [mutation_prob]*num_generations
-        CXPB_LIST = [crossover_prob]*num_generations
-
-    return MUTPB_LIST, CXPB_LIST
-
-def online_expected_diversity(current_generation = int, max_generations = int, current_diversity = float, old_mut_param = dict):
-    """Model the relationship between the number of generations and diversity."""
-    
-    generation = max(0, min(current_generation, max_generations))
-    midpoint = max_generations / 2.0 # Adjusts the parameters of the sigmoidal function
-    steepness = 0.1  # You can adjust this value to control the steepness of the curve
-    expected_diversity = 1 / (1 + math.exp(-steepness * (generation - midpoint))) # Calculates diversity using a sigmoidal function
-
-    if current_generation == 0:
-        std_dev = random.uniform(0.05, 0.3)
-        eta = random.randint(10, 20)
-    else:
-        if current_diversity < expected_diversity:
-            std_dev = abs(old_mut_param["std_dev"][current_generation-1] + old_mut_param["std_dev"][current_generation-1]*(current_diversity/expected_diversity))
-            eta = abs(old_mut_param["eta"][current_generation-1] + old_mut_param["eta"][current_generation-1]*(current_diversity/expected_diversity))
-        elif current_diversity > expected_diversity:
-            std_dev = abs(old_mut_param["std_dev"][current_generation-1] - old_mut_param["std_dev"][current_generation-1]*(expected_diversity/current_diversity))
-            eta = abs(old_mut_param["eta"][current_generation-1] - old_mut_param["eta"][current_generation-1]*(expected_diversity/current_diversity))
-        elif current_diversity == expected_diversity:
-            std_dev = old_mut_param["std_dev"][current_generation-1]
-            eta = old_mut_param["eta"][current_generation-1]
-
-    old_mut_param["std_dev"].append(std_dev); old_mut_param["eta"].append(eta)
-
-    return std_dev, eta, old_mut_param
-
-
-# #####################################################################################
-# #################################### Graphs #########################################
-# #####################################################################################
-
-def sensibility(individual = list, fitness_fn = None, increment = None, min_values = list, max_values = list):
-    """Calculate the fitness of an individual for each iteration, where one variable is incremented by a given value within the range of min and max values.
-    If variable is integer, it will increment by 1 instead of a float value.
-    """
-    dict = {"nvar":[],"value":[],"fit":[]}
-
-    if isinstance(increment, float): step = [ increment for _ in range(len(min_values))]
-    elif isinstance(increment, list): step = increment
-
-    for i in range(len(individual)):
-        settings.log.info('Iteração: {} de {}'.format(i, len(individual)))
-        for new_value in np.arange(min_values[i], max_values[i], step[i]):
-            new_individual = individual.copy()
-            if isinstance(new_individual[i], int):
-                new_value = int(new_value)
-            new_individual[i] = new_value
-            dict["nvar"].append(i)
-            dict["value"].append(new_value)
-            dict["fit"].append(fitness_fn(new_individual))
-
-    df = pd.DataFrame(dict)
-    now = datetime.now()
-    dt_string = now.strftime("%d-%m-%Y_%H-%M")
-    string = 'Resultados/Sensibility_' + str(dt_string) + '.xlsx'
-    df.to_excel(string, index=False)
-
-    return print(pd.DataFrame(dict))
-
-def create_plotfit(num_generations = int, values_gen = None):
-    """Plot the fit and metrics values over the number of generations."""
-   
-    fig, (ax1, ax2, ax3) = plt.subplots(3, 1)
-    fig.subplots_adjust(hspace=0.5)
-
-    ax1.plot(values_gen["best_fit"])
-    ax1.set_xlim(0, num_generations - 1)
-    ax1.set_title('BestFit x Iterations')
-    ax1.set_ylabel('Best Fitness')
-    ax1.grid(True)
-
-    ax2.plot(values_gen["avg_fit"], alpha = 0.3, color = 'red',linestyle = "--")
-    ax2.set_xlim(0, num_generations - 1)
-    ax2.set_title('AvgFit x Iterations')
-    ax2.set_ylabel('Average Fitness')
-    ax2.grid(True)
-
-    ax3.plot(values_gen["metrics"])
-    ax3.set_xlim(0, num_generations - 1)
-    ax3.set_title('Population Diversity x Iterations')
-    ax3.set_ylabel('Diversity Metric')
-    ax3.set_xlabel('Iterations')
-    ax3.grid(True)
-
-    plt.show()
-
-def create_boxplots(out = None, min_values = list, max_values = list):
-    """Boxplot of all values used in the optimization for each variable."""
-
-    history = out["history_valid"]    
-    data = []; aux = []; aux2 = []
-  
-    for i in range(len(history["ind"][0])):
-        for j in range(len(history["ind"])):
-            aux.append(history["ind"][j][i])
-        aux2.append(aux)
-        aux = []
-
-    data = list(map(lambda *x: list(x), *aux2))
-    data_aux = data
-
-    for i in range(len(data)):
-        for j in range(len(min_values)):
-            data_aux[i][j] = ((data[i][j] - min_values[j])/(max_values[j] - min_values[j]))
-
-    df = pd.DataFrame(data_aux)
- 
-    plt.boxplot(df, vert=True)
-    plt.title('Dispersion of values')
-    plt.xlabel('Variables')
-    plt.grid(True)
-    
-    plt.show()
-
-def create_boxplots_import_xlsx(path = None):
-    """Boxplot of all values used in the optimization for each variable."""
-
-    df = pd.read_excel(path)
-    del df["gen"]
-    del df["fit"]
-    del df["score"]
- 
-    plt.boxplot(df, vert=True)
-    plt.title('Dispersion of values')
-    plt.xlabel('Variables')
-    plt.grid(True)
-    
-    plt.show()
-
-def create_boxplots_por_gen_import_xlsx(path = None, min_values = list, max_values = list, generation = int):
-    """Boxplot of all values used in the generation for each variable."""
-
-    df = pd.read_excel(path)
-    del df["fit"]
-    del df["score"]
-    
-    filter = df['gen'] == generation
-    df_aux = df[filter]
-    del df_aux["gen"]
-
-    lista = df_aux.values.tolist()
-    aux = [ [ 0 for _ in range(len(lista[0]))] for _ in range(len(lista)) ]
-    for i in range(len(lista)):
-        for j in range(len(lista[0])):
-            aux[i][j] = (lista[i][j]-min_values[j])/(max_values[j]-min_values[j])
-
-    df = pd.DataFrame(aux)
-
-    plt.boxplot(df, vert=True)
-    plt.title('Value dispersion in generation ' + str(generation))
-    plt.xlabel('Variables')
-    plt.ylabel('Normalized data')
-    plt.grid(True)
-    
-    plt.show()
-
-def parallel_coordinates(out = None, lb = list, ub = list):
-    """Create a parallel coordinates graph of the population history."""
-    
-    history = out["history_valid"]
-    lista = list(history["score"])
-
-    num_ind = len(history["ind"])
-    num_var = len(history["ind"][0])
-    
-    data = []; aux = []; aux2 = []
-  
-    for i in range(num_var):
-        for j in range(num_ind):
-            aux.append(history["ind"][j][i])
-        aux2.append(aux)
-        aux = []
-
-    data = list(map(lambda *x: list(x), *aux2))
-    df = pd.DataFrame(data)
-    df['Score'] = lista
-    
-    fig = px.parallel_coordinates(df, color="score", dimensions=df.columns,title="Parallel Coordinates Plot")
-
-    # Ajustar os limites dos eixos
-    for i, dim in enumerate(df.columns[:-1]):
-        fig.update_layout(yaxis_range=[lb[i], ub[i]], row=i + 1)
-
-    fig.show()
-
-def parallel_coordinates_import_xlsx(path = None, classe = None, lb = list, ub = list):
-    """Create a parallel coordinates graph of the population history."""
-    
-    df = pd.read_excel(path)
-    del df["gen"]
-    del df["fit"]
-
-    micro = ['c1', 'chord_ratio2','b1','span_ratio2','iw','nperfilw1','nperfilw2','zwGround','xCG','vh', 'ih','nperfilh','motorindex']
-    regular = ['b1', 'span_ratio_2', 'span_ratio_b3', 'c1', 'chord_ratio_c2', 'chord_ratio_c3', 'nperfilw1', 'nperfilw2', 'nperfilw3', 'iw', 'zwground', 'xCG', 'Vh', 'ARh', 'nperfilh', 'lt', 'it', 'xTDP', 'AtivaProfundor', 'motorIndex']
-
-    if classe != None:
-        if isinstance(classe,list):
-            for i in range(df.shape[1]-1):
-                df = df.rename({df.columns[i]: classe[i]}, axis='columns')
-        else:
-            if classe == "micro":
-                nomes = micro
-            else:
-                nomes = regular
-            for i in range(df.shape[1]-1):
-                df = df.rename({df.columns[i]: nomes[i]}, axis='columns')
-   
-    fig = px.parallel_coordinates(df, color="score", dimensions=df.columns,title="Parallel Coordinates Plot")
-
-    # Ajustar os limites dos eixos
-    for i, dim in enumerate(df.columns[:-1]):
-        fig.update_layout(yaxis_range=[lb[i], ub[i]], row=i + 1)
-    
-    fig.show()
-
-def parallel_coordinates_per_gen_import_xlsx(path = None, classe = None, lb = list, ub = list, generation = int):
-    """Create a parallel coordinates graph of the population history."""
-    
-    df_aux = pd.read_excel(path)
-   
-    filter = df_aux['gen'] == generation
-    df = df_aux[filter]
-
-    del df["gen"]
-    del df["fit"]
-
-    micro = ['c1', 'chord_ratio2','b1','span_ratio2','iw','nperfilw1','nperfilw2','zwGround','xCG','vh', 'ih','nperfilh','motorindex']
-    regular = ['b1', 'span_ratio_2', 'span_ratio_b3', 'c1', 'chord_ratio_c2', 'chord_ratio_c3', 'nperfilw1', 'nperfilw2', 'nperfilw3', 'iw', 'zwground', 'xCG', 'Vh', 'ARh', 'nperfilh', 'lt', 'it', 'xTDP', 'AtivaProfundor', 'motorIndex']
-
-    if classe != None:
-        if isinstance(classe,list):
-            for i in range(df.shape[1]-1):
-                df = df.rename({df.columns[i]: classe[i]}, axis='columns')
-        else:
-            if classe == "micro":
-                nomes = micro
-            else:
-                nomes = regular
-            for i in range(df.shape[1]-1):
-                df = df.rename({df.columns[i]: nomes[i]}, axis='columns')
-   
-    fig = px.parallel_coordinates(df, color="score", dimensions=df.columns,title="Parallel Coordinates Plot")
-
-    # Ajustar os limites dos eixos
-    for i, dim in enumerate(df.columns[:-1]):
-        fig.update_layout(yaxis_range=[lb[i], ub[i]], row=i + 1)
-
-    fig.show()
-
-
-def export_excell(out):
-    """Create a parallel coordinates graph of the population history."""
-    
-    history = out["history"]
-    lista = list(history["gen"])
-    lista2 = list(history["fit"])
-    lista3 = list(history["score"])
-
-    num_ind = len(history["ind"])
-    num_var = len(history["ind"][0])
-    
-    data = []; aux = []; aux2 = []
-  
-    for i in range(num_var):
-        for j in range(num_ind):
-            aux.append(history["ind"][j][i])
-        aux2.append(aux)
-        aux = []
-
-    data = list(map(lambda *x: list(x), *aux2))
-    df = pd.DataFrame(data)
-
-    df['gen'] = lista
-    df['fit'] = lista2
-    df['score'] = lista3
-
-    # Check whether the specified path exists or not
-    path = "Resultados"
-    pathExist = os.path.exists(path)
-    if not pathExist:
-        os.makedirs(path)   
-
-    now = datetime.now()
-    dt_string = now.strftime("%d-%m-%Y_%H-%M")
-    string = 'Resultados/Results_' + str(dt_string) + '.xlsx'
-
-    df.to_excel(string, index=False)
-
-# #####################################################################################
-# ##################################### Tests #########################################
-# #####################################################################################
-
-def run_n_times(selection1, crossover1, mutation1, n_threads1,
-    min_values1, max_values1, num_variables1, num_generations1, elite_count1,
-    fitness_fn1, classe1, n):
-
-    fit = []
-    best_fit = []
-    
-    for _ in range(0,n):
-
-        out = optimize(selection = selection1, crossover = crossover1, mutation = mutation1, n_threads = n_threads1,
-        min_values = min_values1, max_values = max_values1, num_variables = num_variables1, num_generations = num_generations1, elite_count = elite_count1,
-        fitness_fn = fitness_fn1, classe = classe1, plotfit=False)
-
-        fit.append(out["best_individual"]["fit"])
-        best_fit.append(min(out["best_individual"]["fit"]))
-    
-    aux = list(map(lambda *x: list(x), *fit))
-    df = pd.DataFrame(aux)
-    df2 = pd.DataFrame(best_fit)
-
-    fig, (ax1, ax2) = plt.subplots(1, 2)
-    fig.subplots_adjust(hspace=0.5)
-
-    ax1.boxplot(df, vert=True)
-    ax1.set_title('Fitness for '+ str(n) +' Runs')
-    ax1.set_ylabel('Fitness')
-    ax1.grid(True)
-
-    ax2.boxplot(df2, vert=True)
-    ax2.set_title('Best Fitness for All Runs')
-    ax2.grid(True)
-
-    plt.show()
