+++ src/AeroGA/Operators_Crossover.py
@@ -1,0 +1,63 @@
+import random
+
+def arithmetic_crossover(parent1 = list, parent2 = list, min_values = list, max_values = list, alpha = 0.05):
+    """Apply arithmetic crossover to produce two offspring."""
+    offspring1 = []
+    offspring2 = []
+    for i in range(len(parent1)):
+        offspring1.append(round(alpha*parent1[i] + (1-alpha)*parent2[i],4))
+        offspring2.append(round(alpha*parent2[i] + (1-alpha)*parent1[i],4))
+        if not isinstance(parent1[i], float):
+            offspring1[i] = int(offspring1[i])
+            offspring2[i] = int(offspring2[i])
+
+        # Ensure the offspring stay within the bounds
+        offspring1[i] = min(max(offspring1[i], min_values[i]), max_values[i])
+        offspring2[i] = min(max(offspring2[i], min_values[i]), max_values[i])
+
+    return offspring1, offspring2
+
+def SBX_crossover(parent1 = list, parent2 = list, min_values = list, max_values = list, eta=0.5):
+    """Apply SBX crossover to produce two offspring."""
+    # Generate a random number for each variable
+    u = random.uniform(0, 1)
+    
+    # Calculate the beta value for each variable
+    beta = 0
+    if u <= 0.5:
+        beta = (2 * u) ** (1 / (eta + 1))
+    else:
+        beta = (1 / (2 * (1 - u))) ** (1 / (eta + 1))
+    
+    # Calculate the offspring for each variable
+    offspring1 = []
+    offspring2 = []
+    for i in range(len(parent1)):
+        offspring1.append(round(0.5 * ((1 + beta) * parent1[i] + (1 - beta) * parent2[i]),4))
+        offspring2.append(round(0.5 * ((1 - beta) * parent1[i] + (1 + beta) * parent2[i]),4))
+        
+        # Ensure the offspring stay within the bounds
+        offspring1[i] = min(max(offspring1[i], min_values[i]), max_values[i])
+        offspring2[i] = min(max(offspring2[i], min_values[i]), max_values[i])
+        
+        # Round integer values to the nearest integer
+        if isinstance(parent1[i], int):
+            offspring1[i] = round(offspring1[i])
+            offspring2[i] = round(offspring2[i])
+    
+    return offspring1, offspring2
+
+def crossover_1pt(parent1 = list, parent2 = list):
+    """Apply 1 Point crossover to produce two offspring."""
+    n = len(parent1)
+    cxpoint = random.randint(1, n-1)
+    offspring1 = parent1[:cxpoint] + parent2[cxpoint:]
+    offspring2 = parent2[:cxpoint] + parent1[cxpoint:]
+    return offspring1, offspring2
+
+def crossover_2pt(parent1 = list, parent2 = list):
+    """Apply 2 Point crossover to produce two offspring."""
+    size = len(parent1)
+    cxpoint1, cxpoint2 = sorted(random.sample(range(size), 2))
+    parent1[cxpoint1:cxpoint2], parent2[cxpoint1:cxpoint2] = parent2[cxpoint1:cxpoint2], parent1[cxpoint1:cxpoint2]
+    return parent1, parent2
